Skyblock Core de para olaylarını kullanmak için provider inşa edildi.

+++ src/main/java/com/sallamadm/skyblockeco/EconomyAPI.java
@@ -18,27 +18,27 @@ public class EconomyAPI {
     }
 
 
-    public static long getBalance(UUID playerUuid) {
-        if (!ready()) return 0L;
+    public static double getBalance(UUID playerUuid) {
+        if (!ready()) return 0.0;
         return eco().getDataManager().getBalance(playerUuid);
     }
 
-    public static boolean setBalance(UUID playerUuid, long amount) {
+    public static boolean setBalance(UUID playerUuid, double amount) {
         if (!ready()) return false;
         return eco().getDataManager().setBalance(playerUuid, amount);
     }
 
-    public static void addBalance(UUID playerUuid, long amount) {
+    public static void addBalance(UUID playerUuid, double amount) {
         if (!ready()) return;
         eco().getDataManager().addBalanceAsync(playerUuid, amount);
     }
 
-    public static boolean removeBalance(UUID playerUuid, long amount) {
+    public static boolean removeBalance(UUID playerUuid, double amount) {
         if (!ready()) return false;
         return eco().getDataManager().removeBalance(playerUuid, amount);
     }
 
-    public static boolean hasBalance(UUID playerUuid, long amount) {
+    public static boolean hasBalance(UUID playerUuid, double amount) {
         if (!ready()) return false;
         return eco().getDataManager().hasBalance(playerUuid, amount);
     }
